requirements.txt
Flask
google-generativeai
Werkzeug
pymongo
gunicorn
dnspython
